+++ docs/source/interpretations.rst
@@ -24,11 +24,3 @@ Monte Carlo
     :undoc-members:
     :show-inheritance:
     :member-order: bysource
-
-Memoize
------------
-.. automodule:: funsor.memoize
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :member-order: bysource
