+++ docs/source/ops.rst
@@ -12,15 +12,7 @@ Operation classes
 Builtin operations
 ------------------
 .. automodule:: funsor.ops.builtin
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :member-order: bysource
 
 Array operations
 ----------------
 .. automodule:: funsor.ops.array
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :member-order: bysource
